Yonas.phase three (#3)
* feat(Backend-Yonas) create Identity files

* feat(Backend-Yonas) Identity layer

* feat(Backend-Yonas) auth-controller

* feat(Backend-Yonas) Authorization

* fix(Backend-Yonas) model refactor

* feat(Backend-Yonas) authorization

* feat(Backend-Yonas) checklist auth

* feat(Backend-Yonas) usertask tests

* feat(Backend-Yonas) identity tests

+++ TaskManagementSystem.Application/Features/CheckList/CQRS/Handlers/Commands/DeleteCheckListCommandHandler.cs
@@ -1,5 +1,6 @@
 using AutoMapper;
 using MediatR;
+using TaskManagementSystem.Application.Contracts.Identity;
 using TaskManagementSystem.Application.Contracts.Persistence;
 using TaskManagementSystem.Application.Exceptions;
 using TaskManagementSystem.Application.Features.CheckList.CQRS.Requests.Commands;
@@ -13,9 +14,12 @@ public class DeleteCheckListCommandHandler : IRequestHandler<DeleteCheckListComm
 
     private readonly IMapper _mapper;
 
-    public DeleteCheckListCommandHandler(IUnitOfWork unitOfWork, IMapper mapper)
+    private readonly IAuthorizationService _authService;
+
+    public DeleteCheckListCommandHandler(IUnitOfWork unitOfWork, IMapper mapper, IAuthorizationService authorizationService)
     {
         _unitOfWork = unitOfWork;
+        _authService = authorizationService;
         _mapper = mapper;
     }
 
@@ -24,12 +28,18 @@ public class DeleteCheckListCommandHandler : IRequestHandler<DeleteCheckListComm
 
         var validator = new DeleteCheckListDtoValidator(_unitOfWork.CheckListRepository);
         var validationResult = await validator.ValidateAsync(request.deleteCheckListDto);
-
              
         if (validationResult.IsValid == false)
             throw new ValidationException(validationResult);
-    
+
         var CheckList = await _unitOfWork.CheckListRepository.Get(request.deleteCheckListDto.Id);
+        var UserTask = await _unitOfWork.UserTaskRepository.Get(CheckList.UserTaskId);
+        
+        var result = await _authService.UserTaskBelongsToUser(UserTask, request.UserId);
+
+        if(result == false)
+            throw new UnauthorizedException("Task doesn't belong to you");
+    
         await _unitOfWork.CheckListRepository.Delete(CheckList);
 
         if (await _unitOfWork.Save() < 0)
